chore: move from packages.config to PackageReference
For a couple of years .NET projects can use `PackageReference` instead of `packages.config` to list dependencies.

This has several immediate benefits. The main advantage is that it simplifies the project files and removes redundancy, since packages only have to be listed in one place, and transitive dependencies no longer need to be listed.

Reference: https://learn.microsoft.com/en-us/nuget/consume-packages/migrate-packages-config-to-package-reference

This is also a step towards migrating the project files to the newer "sdk-style" format.